правки module
